+++ AGENTS.md
@@ -827,3 +827,4 @@
 - Restored fire icon in like buttons and guarded legacy search click handler (PR like-icon-restore)
 - Consolidated DOMContentLoaded handlers from comment.js, forum_editor.js, enhanced-ui.js, store.js, league.js and feed.js into main.js; modules expose init functions (PR domcontent-modules-consolidation).
 - Guarded reaction button initialization to prevent TypeError on pages without .btn-reaction (PR feed-reaction-null-check).
+- Feed routes split into subpackage crunevo/routes/feed with views.py and api.py; feed_routes.py now re-exports for compatibility (PR feed-subpackage).
